web/requirements.txt
asgiref==3.8.1
cassandra-driver==3.29.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.7
django-cassandra-engine==1.9.0
geomet==0.2.1.post1
idna==3.10
PyYAML==6.0.2
requests==2.32.3
scylla-driver==3.28.2
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
